feat(retrievers): rename the registry modules
