+++ scripts/evaluate_retriever.py
@@ -1,15 +1,13 @@
-import json
 from typing import Annotated, cast
 
+import huggingface_hub
 import typer
 from datasets import Dataset, load_dataset
 from dotenv import load_dotenv
 from vidore_benchmark.evaluation.evaluate import evaluate_dataset
-from vidore_benchmark.retrievers.utils.initialize_retrievers import load_vision_retriever_from_registry
+from vidore_benchmark.retrievers.utils.load_retriever import load_vision_retriever_from_registry
 from vidore_benchmark.utils.constants import OUTPUT_DIR
-from vidore_benchmark.evaluation.evaluate import evaluate_dataset
 from vidore_benchmark.utils.log_utils import log_metrics
-import huggingface_hub
 
 load_dotenv(override=True)
