+++ src/vidore_benchmark/retrievers/__init__.py
@@ -1,8 +1,8 @@
-from .colpali_retriever import ColPaliRetriever
 from .bge_m3_retriever import BGEM3Retriever
+from .bm25_retriever import BM25Retriever
+from .colpali_retriever import ColPaliRetriever
+from .dummy_retriever import DummyRetriever
 from .jina_clip_retriever import JinaClipRetriever
 from .nomic_retriever import NomicVisionRetriever
-from .dummy_retriever import DummyRetriever
 from .siglip_retriever import SigLIPRetriever
 from .vision_retriever import VisionRetriever
-from .bm25_retriever import BM25Retriever
