+++ src/vidore_benchmark/retrievers/bge_m3_retriever.py
@@ -5,7 +5,7 @@ from FlagEmbedding import BGEM3FlagModel
 from PIL import Image
 from tqdm import tqdm
 
-from vidore_benchmark.retrievers.utils.register_models import register_vision_retriever
+from vidore_benchmark.retrievers.utils.register_retriever import register_vision_retriever
 from vidore_benchmark.retrievers.vision_retriever import VisionRetriever
 from vidore_benchmark.utils.iter_utils import batched
 from vidore_benchmark.utils.torch_utils import get_torch_device
