+++ src/vidore_benchmark/retrievers/jina_clip_retriever.py
@@ -7,7 +7,7 @@ from PIL import Image
 from tqdm import tqdm
 from transformers import AutoModel
 
-from vidore_benchmark.retrievers.utils.register_models import register_vision_retriever
+from vidore_benchmark.retrievers.utils.register_retriever import register_vision_retriever
 from vidore_benchmark.retrievers.vision_retriever import VisionRetriever
 from vidore_benchmark.utils.iter_utils import batched
 from vidore_benchmark.utils.torch_utils import get_torch_device
