+++ src/vidore_benchmark/retrievers/siglip_retriever.py
@@ -6,7 +6,7 @@ from torch._tensor import Tensor
 from tqdm import tqdm
 from transformers import AutoModel, AutoProcessor
 
-from vidore_benchmark.retrievers.utils.register_models import register_vision_retriever
+from vidore_benchmark.retrievers.utils.register_retriever import register_vision_retriever
 from vidore_benchmark.retrievers.vision_retriever import VisionRetriever
 from vidore_benchmark.utils.iter_utils import batched
 from vidore_benchmark.utils.torch_utils import get_torch_device
