+++ src/vidore_benchmark/retrievers/utils/load_retriever.py
@@ -1,10 +1,15 @@
-from vidore_benchmark.retrievers.utils.register_models import VISION_RETRIEVER_REGISTRY
+from vidore_benchmark.retrievers.utils.register_retriever import VISION_RETRIEVER_REGISTRY
 from vidore_benchmark.retrievers.vision_retriever import VisionRetriever
 
 
 def load_vision_retriever_from_registry(model_name: str, *args, **kwargs) -> VisionRetriever:
     """
-    Create a vision retriever instance based on the model name.
+    Create a vision retriever instance based on the associated model name.
+
+    To name an instance of VisionRetriever, use the following decorator:
+    >>> @register_vision_retriever("my_vision_retriever")
+    >>> class MyVisionRetriever(VisionRetriever):
+    >>> ...
     """
 
     if model_name in VISION_RETRIEVER_REGISTRY:
