+++ src/vidore_benchmark/retrievers/utils/register_retriever.py
